Introduces izettle-kotlin library
This introduces a new izettle-kotlin library into the toolbox. The motivation for this library is simply to make programming in Kotlin at iZettle easier and more enjoyable, and to reduce boilerplate and ceremony (a primary goal of Kotlin)

Initially, I have provided some functions that I think will make our Kotlin code more "Kotlin-like", by encouraging chains of calls and not if statements. We are already using the whenNotNull function in https://github.com/iZettle/online-payment and it would be much better to have it in a library like this one.

```
// Introduces a whenNoneNull() global function that acts like a multi-variable ?.let()
// Supports between 2 and 9 arguments of different or same types

whenNoneNull(nullableName, nullableAge) { name, age ->
    println("Block was called with $name, $age, guaranteed to be non null")
} ?: throw Exception()

// Adds a mapper between Java Optional to a nullable Kotlin type,
// to cut down on boilerplate in Kotlin code where we get Optionals
// from legacy Java frameworks or code

val result = optional.asNullable()?.let { "has value $it" } ?: "empty"
```

Also: updates basepom and makes required checkstyle changes.