Java toolbox: Java 11 adoptions
Turns out that the currency format for different locales has changed in
newer versions of java. Most significantly, non breaking space and
unicode "minus sign" is sometimes used instead of horizontal space and
hyphen.

We choose to live with this change, as we don't want to add custom
configuration that's costly to maintain.

But, as a consequence of this, the parser cannot parse strings with
normal space or hyphen, should those be the characters preferred by the
formatter. So, to not make this too dangerous, logic has been changed in
the parser method, so that normal space and hyphen won't cause a
problem.